Split amp-accordion component files
