+++ extensions/amp-accordion/1.0/component.js
@@ -33,7 +33,7 @@ import {
   useRef,
   useState,
 } from '../../../src/preact';
-import {useStyles} from './accordion.jss';
+import {useStyles} from './component.jss';
 
 const AccordionContext = Preact.createContext(
   /** @type {AccordionDef.AccordionContext} */ ({})
