+++ extensions/amp-accordion/1.0/component.type.js
