+++ extensions/amp-accordion/1.0/storybook/Basic.js
@@ -20,7 +20,7 @@ import {
   AccordionContent,
   AccordionHeader,
   AccordionSection,
-} from '../accordion';
+} from '../component';
 import {boolean, withKnobs} from '@storybook/addon-knobs';
 import {withA11y} from '@storybook/addon-a11y';
